+++ Source_Files/RenderMain/ImageLoader_Shared.cpp
@@ -51,7 +51,7 @@
 #endif
 
 #include <cmath>
-static inline float log2(int x) { return std::log(x) / std::log(2.0); };
+static inline float log2(int x) { return std::log((float) x) / std::log(2.0); };
 
 #include <stdlib.h>
